Purge the cached information in appropriate places.
When the media of a removable device is removed, the cached
information must be purged so that new information can be retrieved
when other media is inserted. Likewise, when a removable device is
added, if it's got no media, its cache must also be discarded.

Thanks to Simon Hyde for the heads up.